third_party/blink/web_tests/fast/canvas-api/canvas-ImageBitmap-close-expected.txt

Tests that the close method of ImageBitmap does dispose the image data

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS bitmap.width is imgWidth
PASS bitmap.height is imgHeight
PASS bitmap.width is imgWidth
PASS bitmap.height is imgHeight
PASS bitmap.width is 0
PASS bitmap.height is 0
PASS ctx.drawImage(bitmap, 0, 0) threw exception InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The image source is detached.
PASS Apply structured clone to an already closed bitmap is rejected as expected: DataCloneError: Failed to execute 'postMessage' on 'Worker': An ImageBitmap is detached and could not be cloned.
PASS Apply transferring to an already closed bitmap is rejected as expected: DataCloneError: Failed to execute 'postMessage' on 'Worker': An ImageBitmap is detached and could not be cloned.
PASS createImageBitmap from a closed ImageBitmap was rejected. InvalidStateError: Failed to execute 'createImageBitmap' on 'Window': The source image width is 0.
PASS bitmap.width is 0
PASS bitmap.height is 0
PASS bitmap.width is imgWidth
PASS bitmap.height is imgHeight
PASS bitmap.width is 0
PASS bitmap.height is 0
PASS bitmap.width is 0
PASS bitmap.height is 0
PASS successfullyParsed is true

TEST COMPLETE
